feat: new included paths to song resources
